pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gallery-algo-starter"
version = "0.1.0"
description = "Starter framework for gallery layout algorithms"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
